Refactor text validator to base class
